@stassats 2021-01-16
If I optimize it in SBCL will you be content with that?