@phmarek 2021-01-16
Yes, of course that fulfills my immediate needs.

And I didn't benchmark on a range of implementations yet anyway.

Thank you very much!